server/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
asttokens==2.4.1
asyncpg==0.29.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
databases==0.9.0
debugpy==1.8.6
decorator==5.1.1
distro==1.9.0
ecdsa==0.19.0
environs==11.0.0
executing==2.1.0
fastapi==0.115.0
fastapi_cors==0.0.6
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jiter==0.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
marshmallow==3.22.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.1
openai==1.50.2
packaging==24.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.11
python-vlc==3.0.20123
pyzmq==26.2.0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.38.6
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
wcwidth==0.2.13
